refactor(pipeline): dedup the batches-view builder between SpriteGroup and Registry
SpriteGroup.batches and Registry.batches ran the identical run-key →
mesh-list traversal independently. Extract buildBatchQueryView into
batchQuery.ts and have both call it — same behavior, one implementation.
