+++ packages/three-flatland/src/orchestration/registry.ts
@@ -2,11 +2,10 @@ import type { Scene, Texture } from 'three'
 import type { World } from 'koota'
 import type { Sprite2D } from '../sprites/Sprite2D'
 import type { Sprite2DMaterial } from '../materials/Sprite2DMaterial'
-import type { SpriteBatch } from '../pipeline/SpriteBatch'
 import { SpriteGroup } from '../pipeline/SpriteGroup'
-import { BatchMesh, BatchRegistry } from '../ecs/traits'
+import { BatchRegistry } from '../ecs/traits'
 import { getWorldDefaultMaterial, type RegistryData } from '../ecs/batchUtils'
-import { BatchQueryView } from '../pipeline/batchQuery'
+import { buildBatchQueryView, type BatchQueryView } from '../pipeline/batchQuery'
 
 /**
  * Module-global registry key. `Symbol.for` survives double-bundling —
@@ -114,18 +113,7 @@ export class Registry {
    * bookkeeping to drift.
    */
   get batches(): BatchQueryView {
-    const view = new BatchQueryView(this.group.world)
-    const data = this._registryData()
-    if (!data) return view
-    for (const [key, run] of data.runs) {
-      const meshes: SpriteBatch[] = []
-      for (const batchEntity of run.batches) {
-        const mesh = batchEntity.get(BatchMesh)?.mesh
-        if (mesh) meshes.push(mesh)
-      }
-      view.set(key, meshes)
-    }
-    return view
+    return buildBatchQueryView(this.group.world, this._registryData())
   }
 
   /** @internal */
