+++ packages/three-flatland/src/pipeline/SpriteGroup.ts
@@ -13,8 +13,7 @@ import {
   getWorldDefaultMaterial,
   removeMaterialDisposeHooks,
 } from '../ecs/batchUtils'
-import { BatchQueryView } from './batchQuery'
-import type { SpriteBatch } from './SpriteBatch'
+import { buildBatchQueryView, type BatchQueryView } from './batchQuery'
 import {
   _registerBatchSource,
   _unregisterBatchSource,
@@ -574,18 +573,7 @@ export class SpriteGroup extends Group implements WorldProvider {
    * classification query facade (`group.batches.where(IsLitBatch)`).
    */
   get batches(): BatchQueryView {
-    const view = new BatchQueryView(this._world)
-    const registry = this._getRegistry()
-    if (!registry) return view
-    for (const [key, run] of registry.runs) {
-      const meshes: SpriteBatch[] = []
-      for (const batchEntity of run.batches) {
-        const mesh = batchEntity.get(BatchMesh)?.mesh
-        if (mesh) meshes.push(mesh)
-      }
-      view.set(key, meshes)
-    }
-    return view
+    return buildBatchQueryView(this._world, this._getRegistry())
   }
 
   /**
